src/mvde/depth/midas.py
"""MiDaS monocular depth estimation."""

from typing import Optional, Tuple, Union
import numpy as np
import torch
import torch.nn.functional as F
from PIL import Image
import cv2


class MiDaSEstimator:
    """MiDaS-based depth estimator."""
    
    def __init__(
        self,
        model_type: str = "DPT_Large",
        device: str = "cuda",
        optimize: bool = True,
    ):
        """
        Initialize MiDaS depth estimator.
        
        Args:
            model_type: MiDaS model variant 
                ("MiDaS", "MiDaS_small", "DPT_Large", "DPT_Hybrid")
            device: Device to run inference on
            optimize: Whether to optimize model for inference
        """
        self.model_type = model_type
        self.device = device
        self.optimize = optimize
        
        # TODO: Load actual MiDaS model
        # import midas
        # self.model = torch.hub.load("intel-isl/MiDaS", model_type)
        # self.model.to(device)
        # self.model.eval()
        
        # if optimize:
        #     self.model = torch.jit.script(self.model)
        
        # Load transforms
        # midas_transforms = torch.hub.load("intel-isl/MiDaS", "transforms")
        # if model_type in ["DPT_Large", "DPT_Hybrid"]:
        #     self.transform = midas_transforms.dpt_transform
        # else:
        #     self.transform = midas_transforms.small_transform
        
        self.model = None  # Placeholder
        self.transform = None  # Placeholder
    
    def estimate_depth(
        self, 
        image: Union[Image.Image, np.ndarray],
        return_tensor: bool = False,
    ) -> Union[np.ndarray, torch.Tensor]:
        """
        Estimate depth for an image.
        
        Args:
            image: Input image (PIL Image or numpy array)
            return_tensor: Whether to return torch tensor or numpy array
            
        Returns:
            Depth map (same spatial dimensions as input)
        """
        # Convert PIL to numpy if needed
        if isinstance(image, Image.Image):
            img_array = np.array(image)
        else:
            img_array = image
        
        # TODO: Implement actual MiDaS inference
        # input_tensor = self.transform(img_array).to(self.device)
        # 
        # with torch.no_grad():
        #     prediction = self.model(input_tensor)
        #     prediction = F.interpolate(
        #         prediction.unsqueeze(1),
        #         size=img_array.shape[:2],
        #         mode="bicubic",
        #         align_corners=False,
        #     ).squeeze()
        
        # Placeholder - return dummy depth map
        h, w = img_array.shape[:2]
        depth = np.random.random((h, w)).astype(np.float32)
        
        if return_tensor:
            return torch.from_numpy(depth)
        
        return depth
    
    def estimate_depth_batch(
        self, 
        images: list,
        return_tensor: bool = False,
    ) -> Union[list, torch.Tensor]:
        """
        Estimate depth for a batch of images.
        
        Args:
            images: List of images
            return_tensor: Whether to return torch tensor
            
        Returns:
            List of depth maps or batched tensor
        """
        depth_maps = []
        
        for image in images:
            depth = self.estimate_depth(image, return_tensor=True)
            depth_maps.append(depth)
        
        if return_tensor:
            return torch.stack(depth_maps)
        
        return [d.numpy() for d in depth_maps]
    
    def normalize_depth(
        self, 
        depth: Union[np.ndarray, torch.Tensor],
        method: str = "minmax",
    ) -> Union[np.ndarray, torch.Tensor]:
        """
        Normalize depth values.
        
        Args:
            depth: Depth map
            method: Normalization method ("minmax", "zscore")
            
        Returns:
            Normalized depth map
        """
        if method == "minmax":
            depth_min = depth.min()
            depth_max = depth.max()
            return (depth - depth_min) / (depth_max - depth_min + 1e-8)
        elif method == "zscore":
            depth_mean = depth.mean()
            depth_std = depth.std()
            return (depth - depth_mean) / (depth_std + 1e-8)
        else:
            raise ValueError(f"Unknown normalization method: {method}")
    
    def get_object_depth(
        self, 
        depth_map: Union[np.ndarray, torch.Tensor],
        bbox: Tuple[int, int, int, int],
        aggregation: str = "median",
    ) -> float:
        """
        Extract depth value for a bounding box region.
        
        Args:
            depth_map: Full depth map
            bbox: Bounding box (x1, y1, x2, y2)
            aggregation: How to aggregate depth in region ("mean", "median", "min")
            
        Returns:
            Aggregated depth value
        """
        x1, y1, x2, y2 = bbox
        
        # Extract region
        if isinstance(depth_map, torch.Tensor):
            region = depth_map[y1:y2, x1:x2]
            
            if aggregation == "mean":
                return region.mean().item()
            elif aggregation == "median":
                return region.median().item()
            elif aggregation == "min":
                return region.min().item()
        else:
            region = depth_map[y1:y2, x1:x2]
            
            if aggregation == "mean":
                return float(np.mean(region))
            elif aggregation == "median":
                return float(np.median(region))
            elif aggregation == "min":
                return float(np.min(region))
        
        raise ValueError(f"Unknown aggregation method: {aggregation}")
    
    def depth_to_distance(
        self, 
        depth_value: float,
        camera_params: Optional[dict] = None,
    ) -> float:
        """
        Convert depth value to real-world distance.
        
        Args:
            depth_value: Raw depth value from model
            camera_params: Optional camera intrinsics for calibration
            
        Returns:
            Distance in meters (or model units)
        """
        # TODO: Implement proper depth to distance conversion
        # This depends on the specific depth model and camera calibration
        
        # Placeholder - assume depth is already in reasonable units
        return depth_value
