src/mvde/detectors/yolo.py
"""YOLO-based object detection for bounding boxes."""

from typing import List, Tuple, Dict, Any, Optional
import numpy as np
from PIL import Image
import torch


class BoundingBox:
    """Bounding box representation."""
    
    def __init__(self, x1: float, y1: float, x2: float, y2: float, 
                 confidence: float, class_id: int, class_name: str):
        self.x1 = x1
        self.y1 = y1  
        self.x2 = x2
        self.y2 = y2
        self.confidence = confidence
        self.class_id = class_id
        self.class_name = class_name
    
    @property
    def center(self) -> Tuple[float, float]:
        """Get center coordinates."""
        return ((self.x1 + self.x2) / 2, (self.y1 + self.y2) / 2)
    
    @property
    def area(self) -> float:
        """Get bounding box area."""
        return (self.x2 - self.x1) * (self.y2 - self.y1)
    
    def to_dict(self) -> Dict[str, Any]:
        """Convert to dictionary."""
        return {
            "bbox": [self.x1, self.y1, self.x2, self.y2],
            "confidence": self.confidence,
            "class_id": self.class_id,
            "class_name": self.class_name,
            "center": self.center,
            "area": self.area,
        }


class YOLODetector:
    """YOLO-based object detector."""
    
    def __init__(
        self, 
        model_name: str = "yolov8n",
        device: str = "cuda",
        confidence_threshold: float = 0.5,
        nms_threshold: float = 0.45,
    ):
        """
        Initialize YOLO detector.
        
        Args:
            model_name: YOLO model variant
            device: Device to run inference on
            confidence_threshold: Minimum confidence for detections
            nms_threshold: NMS threshold for duplicate removal
        """
        self.model_name = model_name
        self.device = device
        self.confidence_threshold = confidence_threshold
        self.nms_threshold = nms_threshold
        
        # TODO: Load actual YOLO model
        # from ultralytics import YOLO
        # self.model = YOLO(model_name)
        # self.model.to(device)
        
        self.model = None  # Placeholder
        
        # COCO class names (placeholder)
        self.class_names = [
            "person", "bicycle", "car", "motorcycle", "airplane", "bus", "train", 
            "truck", "boat", "traffic light", "fire hydrant", "stop sign",
            # ... add all 80 COCO classes
        ]
    
    def detect(self, image: Image.Image) -> List[BoundingBox]:
        """
        Detect objects in an image.
        
        Args:
            image: PIL Image
            
        Returns:
            List of detected bounding boxes
        """
        # TODO: Implement actual YOLO inference
        # results = self.model(image)
        # return self._parse_results(results)
        
        # Placeholder - return dummy detections
        return [
            BoundingBox(
                x1=100, y1=100, x2=200, y2=200,
                confidence=0.9, class_id=0, class_name="person"
            ),
            BoundingBox(
                x1=300, y1=150, x2=450, y2=300,
                confidence=0.8, class_id=2, class_name="car"
            ),
        ]
    
    def _parse_results(self, results: Any) -> List[BoundingBox]:
        """Parse YOLO results into BoundingBox objects."""
        # TODO: Implement results parsing
        boxes = []
        return boxes
    
    def filter_by_confidence(
        self, 
        boxes: List[BoundingBox], 
        min_confidence: float
    ) -> List[BoundingBox]:
        """Filter boxes by confidence threshold."""
        return [box for box in boxes if box.confidence >= min_confidence]
    
    def filter_by_classes(
        self, 
        boxes: List[BoundingBox], 
        allowed_classes: List[str]
    ) -> List[BoundingBox]:
        """Filter boxes by allowed class names."""
        return [box for box in boxes if box.class_name in allowed_classes]
    
    def non_max_suppression(self, boxes: List[BoundingBox]) -> List[BoundingBox]:
        """Apply non-maximum suppression to remove overlapping boxes."""
        # TODO: Implement NMS algorithm
        return boxes  # Placeholder
